--- arcjet-guard/src/fetch.test.ts
@@ -1,45 +1,0 @@
-import assert from "node:assert/strict";
-import { describe, test } from "node:test";
-
-import {
-  launchArcjet,
-  createTransport,
-  tokenBucket,
-  fixedWindow,
-  slidingWindow,
-  detectPromptInjection,
-  experimental_moderateContent,
-  localDetectSensitiveInfo,
-  defineCustomRule,
-  launchArcjetWithTransport,
-} from "./fetch.ts";
-
-describe("fetch entrypoint", () => {
-  test("launchArcjet is exported as a function", () => {
-    assert.equal(typeof launchArcjet, "function");
-  });
-
-  test("createTransport is re-exported", () => {
-    assert.equal(typeof createTransport, "function");
-  });
-
-  test("rule factories are re-exported", () => {
-    assert.equal(typeof tokenBucket, "function");
-    assert.equal(typeof fixedWindow, "function");
-    assert.equal(typeof slidingWindow, "function");
-    assert.equal(typeof detectPromptInjection, "function");
-    assert.equal(typeof experimental_moderateContent, "function");
-    assert.equal(typeof localDetectSensitiveInfo, "function");
-    assert.equal(typeof defineCustomRule, "function");
-  });
-
-  test("launchArcjetWithTransport is re-exported", () => {
-    assert.equal(typeof launchArcjetWithTransport, "function");
-  });
-
-  test("launchArcjet returns an object with .guard()", () => {
-    const arcjet = launchArcjet({ key: "ajkey_test" });
-
-    assert.equal(typeof arcjet.guard, "function");
-  });
-});
